+++ packages/lumx-core/src/js/constants/index.ts
@@ -2,3 +2,4 @@ export * from './keycodes';
 export * from './components';
 export * from './enums';
 export * from './className';
+export * from './browser';