+++ packages/lumx-react/src/components/text-field/TextField.test.tsx
@@ -13,13 +13,13 @@ import {
 import partition from 'lodash/partition';
 import userEvent from '@testing-library/user-event';
 
-import { isFocusVisible } from '@lumx/react/utils/browser/isFocusVisible';
+import { isFocusVisible } from '@lumx/core/js/utils/browser/isFocusVisible';
 import { createRef } from 'react';
 import { TextField, TextFieldProps } from './TextField';
 
 const CLASSNAME = TextField.className as string;
 
-vi.mock('@lumx/react/utils/browser/isFocusVisible');
+vi.mock('@lumx/core/js/utils/browser/isFocusVisible');
 
 /**
  * Mounts the component and returns common DOM elements / data needed in multiple tests further down.