+++ packages/lumx-vue/src/index.ts
@@ -23,3 +23,4 @@ export * from './components/table';
 export * from './components/text';
 export * from './components/thumbnail';
 export * from './components/toolbar';
+export * from './components/tooltip';