+++ packages/site-demo/content/product/components/tooltip/react/demo2.tsx
@@ -1,22 +1,23 @@
 import { mdiPrinter } from '@lumx/icons';
+import { classNames } from '@lumx/core/js/utils';
 import { Icon, Placement, Tooltip } from '@lumx/react';
 
 export default () => (
     <>
         <Tooltip label="Print" placement={Placement.LEFT} forceOpen>
-            <Icon hasShape icon={mdiPrinter} />
+            <Icon className={classNames.margin('all', 'huge')} hasShape icon={mdiPrinter} />
         </Tooltip>
 
         <Tooltip label="Print" placement={Placement.BOTTOM} forceOpen>
-            <Icon hasShape icon={mdiPrinter} />
+            <Icon className={classNames.margin('all', 'huge')} hasShape icon={mdiPrinter} />
         </Tooltip>
 
         <Tooltip label="Print" placement={Placement.TOP} forceOpen>
-            <Icon hasShape icon={mdiPrinter} />
+            <Icon className={classNames.margin('all', 'huge')} hasShape icon={mdiPrinter} />
         </Tooltip>
 
         <Tooltip label="Print" placement={Placement.RIGHT} forceOpen>
-            <Icon hasShape icon={mdiPrinter} />
+            <Icon className={classNames.margin('all', 'huge')} hasShape icon={mdiPrinter} />
         </Tooltip>
     </>
 );