+++ packages/site-demo/content/product/components/tooltip/vue/demo1.vue
@@ -1,0 +1,10 @@
+<template>
+    <Tooltip label="Print">
+        <Icon :icon="mdiPrinter" :has-shape="true" />
+    </Tooltip>
+</template>
+
+<script setup lang="ts">
+import { mdiPrinter } from '@lumx/icons';
+import { Icon, Tooltip } from '@lumx/vue';
+</script>